+++ library/core/src/test/java/com/google/android/exoplayer2/offline/DownloadManagerTest.java
@@ -583,6 +583,64 @@ public class DownloadManagerTest {
   }
 
   @Test
+  public void addDownload_whilstRemovingWithStopReason_addsStartedDownload() throws Throwable {
+    runOnMainThread(
+        () -> downloadManager.addDownload(createDownloadRequest(ID1), /* stopReason= */ 1234));
+
+    postRemoveRequest(ID1);
+    FakeDownloader downloadRemover = getDownloaderAt(0);
+    downloadRemover.assertRemoveStarted();
+
+    // Re-add the download without a stop reason.
+    postDownloadRequest(ID1);
+
+    downloadRemover.finish();
+
+    FakeDownloader downloader = getDownloaderAt(1);
+    downloader.finish();
+
+    assertDownloadIndexSize(1);
+    // We expect one downloader for the removal, and one for when the download was re-added.
+    assertDownloaderCount(2);
+    // The download has completed, and so is no longer current.
+    assertCurrentDownloadCount(0);
+
+    Download download = postGetDownloadIndex().getDownload(ID1);
+    assertThat(download.state).isEqualTo(Download.STATE_COMPLETED);
+    assertThat(download.stopReason).isEqualTo(0);
+  }
+
+  /** Test for https://github.com/google/ExoPlayer/issues/8419 */
+  @Test
+  public void addDownloadWithStopReason_whilstRemoving_addsStoppedDownload() throws Throwable {
+    postDownloadRequest(ID1);
+    getDownloaderAt(0).finish();
+
+    postRemoveRequest(ID1);
+    FakeDownloader downloadRemover = getDownloaderAt(1);
+    downloadRemover.assertRemoveStarted();
+
+    // Re-add the download with a stop reason.
+    runOnMainThread(
+        () -> downloadManager.addDownload(createDownloadRequest(ID1), /* stopReason= */ 1234));
+
+    downloadRemover.finish();
+
+    assertDownloadIndexSize(1);
+    // We expect one downloader for the initial download, and one for the removal. A downloader
+    // should not be created when the download is re-added, since a stop reason is specified.
+    assertDownloaderCount(2);
+    // The download isn't completed, and is therefore still current.
+    assertCurrentDownloadCount(1);
+
+    List<Download> downloads = postGetCurrentDownloads();
+    Download download = downloads.get(0);
+    assertThat(download.request.id).isEqualTo(ID1);
+    assertThat(download.state).isEqualTo(Download.STATE_STOPPED);
+    assertThat(download.stopReason).isEqualTo(1234);
+  }
+
+  @Test
   public void mergeRequest_removing_becomesRestarting() {
     DownloadRequest downloadRequest = createDownloadRequest(ID1);
     DownloadBuilder downloadBuilder =